Added fractional damage. Each part of radiation suit provides one quarter of protection of overall value.

+++ src/main/java/com/grim3212/mc/pack/industry/config/IndustryConfig.java
@@ -49,7 +49,7 @@ public class IndustryConfig extends GrimConfig {
 	public static float speedModifierMagmaCream;
 	
 	//Uranium Damage - add by SH
-	public static int uraniumDamage;
+	public static float uraniumDamage;
 
 	// Client Only Settings
 	public static boolean showFanParticles;
